K06-broken-auth: Ingress object without TLS cert
Creating an Ingress Object without TLS certificates can be a security concern, as it may lead to potential vulnerabilities and data exposure. While TLS can be used for encryption, TLS also provides a level of authentication by verifying the identity of the server to which a client is connecting. This is done through the server presenting a digital certificate signed by a trusted Certificate Authority (CA). The client can verify the certificate to ensure that it is connecting to the legitimate server and not an imposter. Sadly, I don't have access to any threat research report statistics on this matter, however, this is generally considered a bad practice and should be included in the report, I believe.